release 0.40
The panes for the streams are now properly set up and displayed.
There is a new progress bar showing the overall progress and the current
speed.
Therefore the output of ffmpeg is not displayed anymore if it contains
the information about the fps or the current time.
The user can now abort ffmpeg properly, the gui sends q to the
InputStream of the process.
If the user specifies a already existing file the gui will ask the user
if it can be overwritten. If it does the file will be deleted.